Moved basket logic into separate file
